refactor: Use hook for components logic
